+++ src/index.ts
@@ -1,5 +1,3 @@
-/// <reference path="./types.d.ts" />
-
 import 'core-js/fn/array/flat-map'
 import EventEmitter from 'little-emitter'
 import nanoid       from 'nanoid/non-secure'